Move encoding and array-like to lib/ for better reuse
