+++ lib/array_like.js
