+++ test/array_like_test.js
